+++ src/StateForge.Replication/StateForgeReplicaNode.cs
@@ -6,6 +6,10 @@ namespace StateForge.Replication
 
         public string RootPath { get; set; }
 
+        public string SiteName { get; set; }
+
+        public string Region { get; set; }
+
         public bool Enabled { get; set; }
 
         public StateForgeReplicaNode()