+++ src/StateForge.Replication/StateForgeSiteRole.cs
@@ -1,0 +1,8 @@
+namespace StateForge.Replication
+{
+    public enum StateForgeSiteRole
+    {
+        Primary = 0,
+        Recovery = 1
+    }
+}